# api/openapi.yaml
openapi: 3.0.3
info:
  title: Booking Platform API
  version: 1.0.0
  description: |
    Multi-category service reservation platform.

    Four surfaces, four different middleware chains:

    | Prefix              | Requires                                       |
    |---------------------|------------------------------------------------|
    | `/api/v1/public`    | nothing (an optional bearer token personalizes) |
    | `/api/v1/me`        | a bearer token, scoped by its `sub`             |
    | `/api/v1/org`       | a bearer token **and** `x-organization-id`      |
    | `/api/v1/admin`     | a bearer token whose `platformRole` is `admin`  |

    **Conventions**

    - Money is always an integer in minor units (`priceCents`); 24500 = 245.00 ETB.
    - Times are always RFC3339 **UTC** on the wire.
    - Times of day (opening hours) are `"HH:MM"` strings in the provider's zone.
    - Errors are always `{ "message": string, "code": string }`.

    **Versioning.** Once the mobile app ships, `/api/v1` is additive-only:
    new optional fields are fine; removing, renaming, retyping a field or
    tightening validation is not.

    **Not yet implemented** (build order milestones 8-11): provider discovery
    search, media uploads, reviews, favourites, notifications and payments.
servers:
  - url: http://localhost:8081
    description: Local development
...
security:
  - bearerAuth: []
paths:
...
  /api/v1/admin/categories:
    get:
      tags: [admin]
      summary: List every category, active or not
      responses:
        "200":
          description: All categories.
          content:
            application/json:
              schema:
                type: object
                properties:
                  categories:
                    type: array
                    items: { $ref: "#/components/schemas/Category" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      summary: Create a category
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string }
                slug: { type: string, description: Derived from the name if omitted. }
                icon: { type: string }
                parentId: { type: string, format: uuid, nullable: true }
                sortOrder: { type: integer }
                isActive: { type: boolean, default: true }
      responses:
        "201":
          description: Created.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Category" }
        "409":
          description: "`CONFLICT` — that slug is taken."
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Error" }
...
  /api/v1/admin/providers:
    get:
      tags: [admin]
      summary: List providers across all tenants
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/ProviderStatus" }
        - name: search
          in: query
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Providers, newest first.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      providers:
                        type: array
                        items: { $ref: "#/components/schemas/Provider" }
        "403": { $ref: "#/components/responses/Forbidden" }
    post:
      tags: [admin]
      summary: Onboard a provider directly
      description: For businesses signed up offline; it has no owner until invited.
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateProviderRequest" }
      responses:
        "201":
          description: Created, pending approval.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Provider" }
...
  /api/v1/admin/bookings:
    get:
      tags: [admin]
      summary: List bookings across all tenants
      parameters:
        - name: status
          in: query
          schema: { $ref: "#/components/schemas/BookingStatus" }
        - name: providerId
          in: query
          schema: { type: string, format: uuid }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: A flat triage view, newest first.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      bookings:
                        type: array
                        items: { $ref: "#/components/schemas/AdminBooking" }
        "403": { $ref: "#/components/responses/Forbidden" }
  /api/v1/admin/users:
    get:
      tags: [admin]
      summary: List platform users
      parameters:
        - name: search
          in: query
          description: Matches name or email.
          schema: { type: string }
        - $ref: "#/components/parameters/Limit"
        - $ref: "#/components/parameters/Offset"
      responses:
        "200":
          description: Users, newest first.
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Page"
                  - type: object
                    properties:
                      users:
                        type: array
                        items: { $ref: "#/components/schemas/User" }
        "403": { $ref: "#/components/responses/Forbidden" }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        A 15-minute access token, verified locally on every request. Refresh it
        with `POST /api/v1/auth/refresh`.
  parameters:
...
    Limit:
      name: limit
      in: query
      schema: { type: integer, minimum: 1, maximum: 100, default: 20 }
    Offset:
      name: offset
      in: query
      schema: { type: integer, minimum: 0, default: 0 }
  responses:
...
    Forbidden:
      description: "`FORBIDDEN` — authenticated, but not allowed to do this."
      content:
        application/json:
          schema: { $ref: "#/components/schemas/Error" }
...
  schemas:
    Error:
      type: object
      required: [message, code]
      properties:
        message:
          type: string
          description: Human-readable and safe to show a user.
          example: That slot was just taken
        code:
          type: string
          description: |
            Stable machine-readable code. Clients switch on this, so it is part
            of the contract and additive-only.
          enum:
            - INVALID_CREDENTIALS
            - EMAIL_TAKEN
            - UNAUTHENTICATED
            - FORBIDDEN
            - NOT_A_MEMBER
            - TOKEN_EXPIRED
            - INVALID_TOKEN
            - VALIDATION_ERROR
            - ORG_REQUIRED
            - NOT_FOUND
            - SLOT_TAKEN
            - OUTSIDE_HOURS
            - INVALID_TRANSITION
            - SERVICE_INACTIVE
            - PROVIDER_INACTIVE
            - CONFLICT
            - INTERNAL_ERROR
          example: SLOT_TAKEN
        details:
          type: object
          additionalProperties: { type: string }
          description: |
            Per-field problems, keyed by attribute. Present on validation
            failures only; a client that reads just `message` and `code` is
            unaffected by it.
...
    Page:
      type: object
      properties:
        total: { type: integer, format: int64 }
        limit: { type: integer }
        offset: { type: integer }
    User:
      type: object
      required: [id, name, email, platformRole, emailVerified, createdAt]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        phone: { type: string, nullable: true }
        platformRole:
          type: string
          enum: [user, admin]
        emailVerified: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    Location:
      type: object
      required: [lat, lng]
      properties:
        lat: { type: number, format: double, minimum: -90, maximum: 90 }
        lng: { type: number, format: double, minimum: -180, maximum: 180 }
    ProviderStatus:
      type: string
      enum: [pending, active, suspended]
    BookingMode:
      type: string
      enum: [request, instant]
      description: |
        `instant` skips `pending` and confirms at creation; `request` waits for
        the provider to approve.
    Provider:
      type: object
      required: [id, slug, name, status, timezone, bookingMode]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        phone: { type: string, nullable: true }
        email: { type: string, nullable: true }
        description: { type: string, nullable: true }
        city: { type: string, nullable: true }
        address: { type: string, nullable: true }
        location: { $ref: "#/components/schemas/Location" }
        timezone:
          type: string
          description: IANA zone; all slot maths happens here.
          example: Africa/Addis_Ababa
        logoUrl: { type: string, nullable: true }
        licenseNumber: { type: string, nullable: true }
        status: { $ref: "#/components/schemas/ProviderStatus" }
        ratingAvg: { type: number, format: double }
        ratingCount: { type: integer }
        bookingMode: { $ref: "#/components/schemas/BookingMode" }
        minLeadMinutes:
          type: integer
          description: Slots sooner than this from now are not offered.
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
    CreateProviderRequest:
      type: object
      required: [name]
      properties:
        name: { type: string }
        slug: { type: string, description: Derived from the name if omitted. }
        phone: { type: string }
        email: { type: string, format: email }
        description: { type: string }
        city: { type: string }
        address: { type: string }
        location: { $ref: "#/components/schemas/Location" }
        timezone: { type: string, default: Africa/Addis_Ababa }
        licenseNumber: { type: string }
        bookingMode: { $ref: "#/components/schemas/BookingMode" }
        minLeadMinutes: { type: integer, minimum: 0 }
        categoryIds:
          type: array
          items: { type: string, format: uuid }
...
    Category:
      type: object
      required: [id, slug, name, isActive]
      properties:
        id: { type: string, format: uuid }
        slug: { type: string }
        name: { type: string }
        icon: { type: string, nullable: true }
        parentId: { type: string, format: uuid, nullable: true }
        sortOrder: { type: integer }
        isActive: { type: boolean }
        createdAt: { type: string, format: date-time }
        updatedAt: { type: string, format: date-time }
...
    BookingStatus:
      type: string
      enum:
        - pending
        - confirmed
        - in_progress
        - completed
        - cancelled_by_customer
        - cancelled_by_provider
        - no_show
...
    AdminBooking:
      type: object
      description: The flat cross-tenant triage view.
      properties:
        id: { type: string, format: uuid }
        code: { type: string }
        status: { $ref: "#/components/schemas/BookingStatus" }
        providerId: { type: string, format: uuid }
        providerSlug: { type: string }
        providerName: { type: string }
        serviceName: { type: string }
        priceCents: { type: integer }
        currency: { type: string }
        durationMinutes: { type: integer }
        startsAt: { type: string, format: date-time }
        endsAt: { type: string, format: date-time }
        customerName: { type: string, nullable: true }
        customerEmail: { type: string, nullable: true }
        createdAt: { type: string, format: date-time }
...