# internal/db/models.go
// Code generated by sqlc. DO NOT EDIT.
// versions:
//   sqlc v1.31.1

package db

import (
	"time"

	"github.com/google/uuid"
	"github.com/jackc/pgx/v5/pgtype"
)

type Booking struct {
	ID              uuid.UUID
	Code            string
	ProviderID      uuid.UUID
	ServiceID       *uuid.UUID
	ResourceID      uuid.UUID
	CustomerID      *uuid.UUID
	StartsAt        time.Time
	EndsAt          time.Time
	Status          string
	ServiceName     string
	PriceCents      int32
	Currency        string
	DurationMinutes int32
	Attributes      []byte
	CustomerNote    *string
	CustomerName    *string
	CustomerPhone   *string
	CancelledBy     *string
	CancelReason    *string
	CreatedAt       time.Time
	UpdatedAt       time.Time
}

type BookingOption struct {
	BookingID            uuid.UUID
	OptionID             *uuid.UUID
	Name                 string
	PriceDeltaCents      int32
	DurationDeltaMinutes int32
	SortOrder            int32
}

type BusinessHour struct {
	ID         uuid.UUID
	ProviderID uuid.UUID
	ResourceID *uuid.UUID
	Weekday    int32
	OpensAt    pgtype.Time
	ClosesAt   pgtype.Time
	CreatedAt  time.Time
	UpdatedAt  time.Time
}

type Category struct {
	ID        uuid.UUID
	Slug      string
	Name      string
	Icon      *string
	ParentID  *uuid.UUID
	SortOrder int32
	IsActive  bool
	CreatedAt time.Time
	UpdatedAt time.Time
}

type CategoryAttribute struct {
	ID         uuid.UUID
	CategoryID uuid.UUID
	Key        string
	Label      string
	DataType   string
	Options    []byte
	Required   bool
	AppliesTo  string
	Filterable bool
	SortOrder  int32
	CreatedAt  time.Time
	UpdatedAt  time.Time
}

type Invitation struct {
	ID         uuid.UUID
	ProviderID uuid.UUID
	Email      string
	Role       string
	TokenHash  string
	ExpiresAt  time.Time
	AcceptedAt *time.Time
	CreatedAt  time.Time
}

type Member struct {
	ProviderID uuid.UUID
	UserID     uuid.UUID
	Role       string
	CreatedAt  time.Time
}

type OptionGroup struct {
	ID            uuid.UUID
	ServiceID     uuid.UUID
	Name          string
	SelectionType string
	IsRequired    bool
	MinSelect     int32
	MaxSelect     *int32
	SortOrder     int32
	CreatedAt     time.Time
	UpdatedAt     time.Time
}

type Provider struct {
	ID             uuid.UUID
	Slug           string
	Name           string
	Phone          *string
	Email          *string
	Description    *string
	City           *string
	Address        *string
	Location       interface{}
	Timezone       string
	LogoUrl        *string
	LogoPublicID   *string
	LicenseNumber  *string
	Status         string
	RatingAvg      pgtype.Numeric
	RatingCount    int32
	BookingMode    string
	MinLeadMinutes int32
	CreatedAt      time.Time
	UpdatedAt      time.Time
}

type ProviderCategory struct {
	ProviderID uuid.UUID
	CategoryID uuid.UUID
}

type RefreshToken struct {
	ID         uuid.UUID
	UserID     uuid.UUID
	TokenHash  string
	ClientType string
	ExpiresAt  time.Time
	RevokedAt  *time.Time
	CreatedAt  time.Time
}

type Resource struct {
	ID         uuid.UUID
	ProviderID uuid.UUID
	Name       string
	UserID     *uuid.UUID
	IsActive   bool
	CreatedAt  time.Time
	UpdatedAt  time.Time
}

type ResourceService struct {
	ResourceID uuid.UUID
	ServiceID  uuid.UUID
}

type ScheduleException struct {
	ID         uuid.UUID
	ProviderID uuid.UUID
	ResourceID *uuid.UUID
	Date       time.Time
	IsClosed   bool
	OpensAt    pgtype.Time
	ClosesAt   pgtype.Time
	Reason     *string
	CreatedAt  time.Time
	UpdatedAt  time.Time
}

type Service struct {
	ID              uuid.UUID
	ProviderID      uuid.UUID
	CategoryID      uuid.UUID
	Name            string
	Description     *string
	PriceCents      int32
	Currency        string
	DurationMinutes int32
	BufferMinutes   int32
	Attributes      []byte
	ImageUrl        *string
	ImagePublicID   *string
	IsActive        bool
	CreatedAt       time.Time
	UpdatedAt       time.Time
}

type ServiceOption struct {
	ID                   uuid.UUID
	GroupID              uuid.UUID
	Name                 string
	PriceDeltaCents      int32
	DurationDeltaMinutes int32
	IsActive             bool
	SortOrder            int32
	CreatedAt            time.Time
	UpdatedAt            time.Time
}

type User struct {
	ID            uuid.UUID
	Name          string
	Email         string
	Phone         *string
	PasswordHash  string
	PlatformRole  string
	EmailVerified bool
	FcmToken      *string
	CreatedAt     time.Time
	UpdatedAt     time.Time
}